Add comprehensive condition expression tests with shared base
- Create ConditionExpressionTestBase in test-base using AttributeValue abstraction
- Add 7 new test cases covering previously untested operators:
  * OR operator (||)
  * Not equal (=!=, renders as <>)
  * Greater than (>)
  * Greater/less than or equal in isolation (>=, <=)
  * Complex AND+OR combinations
  * Attribute-to-attribute not-equal comparison
- Refactor aws1/aws2 tests to extend base (eliminates ~400 lines duplication)
- Tests defined once, run on both AWS SDK v1 and v2 implementations

Follows DynamoDBTestBase pattern for shared test logic across implementations.
