+++ dynamodb-local-metadata.json
@@ -1,0 +1,1 @@
+{"installationId":"05976b4e-3775-41f2-897e-8d485e6be679","telemetryEnabled":"true"}
